updated auth slice and auth functions
