+++ src/features/Auth/authSlice.js
@@ -8,6 +8,9 @@ const initialState = {
   photoURL: "",
   isLoggedIn: false,
   error: "",
+  link: "",
+  bio: "",
+  gender: "",
 };
 
 const authSlice = createSlice({
@@ -15,13 +18,17 @@ const authSlice = createSlice({
   initialState,
   reducers: {
     loginAction: (state, action) => {
-      const { displayName, uid, email, photoURL, username } = action.payload;
+      const { displayName, uid, email, photoURL, username, link, bio, gender } =
+        action.payload;
       state.displayName = displayName;
       state.uid = uid;
       state.email = email;
       state.photoURL = photoURL;
       state.isLoggedIn = true;
       state.username = username;
+      state.bio = bio;
+      state.link = link;
+      state.gender = gender;
     },
     logoutUser: (state) => {
       state.displayName = "";
@@ -30,11 +37,17 @@ const authSlice = createSlice({
       state.photoURL = "";
       state.uid = "";
       state.username = "";
+      state.bio = "";
+      state.link = "";
+      state.gender = "";
     },
     updateProfileAction: (state, action) => {
-      const { displayName, photoURL } = action.payload;
+      const { displayName, photoURL, bio, link, gender } = action.payload;
       state.displayName = displayName;
       state.photoURL = photoURL;
+      state.bio = bio;
+      state.gender = gender;
+      state.link = link;
     },
   },
 });
