+++ src/services/index.js
@@ -1,25 +1,56 @@
-import { logout, updateUserProfile } from "./authServices";
+import {
+  logout,
+  loginUser,
+  signupUser,
+  googleSignInHandler,
+  userNameExists,
+} from "./authServices";
+import {
+  singlePostQuery,
+  queryUserCollection,
+  userDocQuerybyId,
+} from "./firebaseQueries";
 import {
   addNewPost,
   editPost,
-  allPostsListener,
   deletePost,
-  latestUserPostsQuery,
   universalSnapshot,
   universalSnapShotDoc,
-  queryUserCollection,
-  userDocQuery,
+  handleLike,
+  isPostLiked,
+  fetchSinglePost,
+  fetchFeedPosts,
+  fetchExplorePosts,
+  fetchUserPosts,
 } from "./post-services";
+import {
+  getUserDetails,
+  fetchUserInfo,
+  updateUserInfo,
+  fetchUserLikedPosts,
+} from "./userServices";
 export {
   logout,
   addNewPost,
   editPost,
-  allPostsListener,
   deletePost,
-  latestUserPostsQuery,
   universalSnapshot,
   queryUserCollection,
-  userDocQuery,
+  userDocQuerybyId,
   universalSnapShotDoc,
-  updateUserProfile,
+  handleLike,
+  isPostLiked,
+  getUserDetails,
+  singlePostQuery,
+  fetchSinglePost,
+  fetchFeedPosts,
+  fetchExplorePosts,
+  fetchUserPosts,
+  fetchUserInfo,
+  updateUserInfo,
+  fetchUserLikedPosts,
+  loginUser,
+  signupUser,
+  googleSignInHandler,
+  userNameExists,
 };
